Move slo schema definition into schema folder
