+++ x-pack/plugins/observability/server/services/slo/slo_repository.ts
@@ -12,9 +12,10 @@ import { pipe } from 'fp-ts/lib/pipeable';
 import { SavedObjectsClientContract } from '@kbn/core-saved-objects-api-server';
 import { SavedObjectsErrorHelpers } from '@kbn/core-saved-objects-utils-server';
 
-import { StoredSLO, SLO, sloSchema } from '../../types/models';
+import { StoredSLO, SLO } from '../../types/models';
 import { SO_SLO_TYPE } from '../../saved_objects';
 import { SLONotFound } from '../../errors';
+import { sloSchema } from '../../types/schema';
 
 export interface Criteria {
   name?: string;
